Split out fake API
